decouple transactions and core endpoints (#14)
