+++ stac_api/resources/item.py
@@ -23,57 +23,6 @@ router = APIRouter()
 NumType = Union[float, int]
 
 
-@router.post(
-    "/collections/{collectionId}/items",
-    response_model=schemas.Item,
-    response_model_exclude_unset=True,
-    response_model_exclude_none=True,
-)
-def create_item_by_id(
-    item: schemas.Item,
-    crud_client: ItemCrudClient = Depends(item_crud_client_factory),
-    base_url: str = Depends(discover_base_url),
-):
-    """Create item (transactions extension)"""
-    row_data = crud_client.create(item)
-    row_data.base_url = base_url
-    return row_data
-
-
-@router.put(
-    "/collections/{collectionId}/items",
-    response_model=schemas.Item,
-    response_model_exclude_unset=True,
-    response_model_exclude_none=True,
-)
-def update_item_by_id(
-    item: schemas.Item,
-    crud_client: ItemCrudClient = Depends(item_crud_client_factory),
-    base_url: str = Depends(discover_base_url),
-):
-    """Update item (transactions extension)"""
-    row_data = crud_client.update(item)
-    row_data.base_url = base_url
-    return row_data
-
-
-@router.delete(
-    "/collections/{collectionId}/items/{itemId}",
-    response_model=schemas.Item,
-    response_model_exclude_unset=True,
-    response_model_exclude_none=True,
-)
-def delete_item_by_id(
-    itemId: str,
-    crud_client: ItemCrudClient = Depends(item_crud_client_factory),
-    base_url: str = Depends(discover_base_url),
-):
-    """Delete item (transactions extension)"""
-    row_data = crud_client.delete(itemId)
-    row_data.base_url = base_url
-    return row_data
-
-
 @router.get(
     "/collections/{collectionId}/items",
     response_model=ItemCollection,
